requirements.txt
flask
flask-cors
tensorflow
pillow
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm